backend/requirements.txt
flask
flask_cors
flask_socketio
eventlet == 0.30.2
langchain
python-dotenv
openai
llama-cpp-python
gunicorn == 20.1.0